Refactor of /lib & /utils structure
