+++ ensawards.org/src/components/atoms/CardBadge.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils/tailwindClassConcatenation";
 export interface CardBadgeProps {
   text: string;
   className: string;
