+++ ensawards.org/src/components/atoms/cards/AppLeaderboardCard.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../../../lib/utils";
+import { cn } from "../../../utils/tailwindClassConcatenation";
 import { shadcnButtonVariants } from "../../shadcn/shadcnButtonStyles";
 import ChevronRightIcon from "../icons/ChevronRightIcon.astro";
 export interface AppLeaderboardCardProps {
