+++ ensawards.org/src/components/atoms/cards/BestPracticeCard.astro
@@ -1,6 +1,6 @@
 ---
-import { cn } from "../../../lib/utils";
-import { getAppSupportColor } from "../../../utils";
+import { cn } from "../../../utils/tailwindClassConcatenation";
+import { getAppSupportColor } from "../../../utils/styles";
 import { shadcnButtonVariants } from "../../shadcn/shadcnButtonStyles";
 export interface BestPracticeCardProps {
   name: string;
