+++ ensawards.org/src/components/atoms/cards/SuggestionCard.astro
@@ -1,6 +1,6 @@
 ---
 import GitHubOutlineIcon from "../../../assets/githubOutlineIcon.svg";
-import { cn } from "../../../lib/utils";
+import { cn } from "../../../utils/tailwindClassConcatenation";
 import { getSuggestionText } from "../../../utils";
 import { PossibleSuggestions } from "../../molecules/contact-form/types";
 import { shadcnButtonVariants } from "../../shadcn/shadcnButtonStyles";
@@ -47,7 +47,7 @@ const { whatsSuggested }: SuggestionCardProps = Astro.props;
   </div>
 </div>
 <script>
-  import { openSuggestionOverlay } from "../../../utils/openSuggestionOverlay";
+  import { openSuggestionOverlay } from "../../../utils/domActions";
 
   class SuggestionFormOpener extends HTMLElement {
     connectedCallback() {
