+++ ensawards.org/src/components/atoms/icons/ChevronRightIcon.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../../../lib/utils";
+import { cn } from "../../../utils/tailwindClassConcatenation";
 import { IconProps } from "./types";
 
 const { className }: IconProps = Astro.props;
