+++ ensawards.org/src/components/molecules/Hero.astro
@@ -1,6 +1,6 @@
 ---
 import HeroImage from "../../assets/mergedHeroImage.png";
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils/tailwindClassConcatenation";
 import { shadcnButtonVariants } from "../shadcn/shadcnButtonStyles";
 ---
