+++ ensawards.org/src/components/molecules/ItemsList.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils/tailwindClassConcatenation";
 import { shadcnButtonVariants } from "../shadcn/shadcnButtonStyles";
 export interface ItemsListProps {
   header: string;
