+++ ensawards.org/src/components/organisms/DataLoadingError.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../../layouts/Layout.astro";
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils/tailwindClassConcatenation";
 import { shadcnButtonVariants } from "../shadcn/shadcnButtonStyles";
 ---
