+++ ensawards.org/src/components/organisms/StaticHeader.astro
@@ -1,6 +1,6 @@
 ---
 import ENSAwardsLogo from "../../assets/ENSAwardsLogo.svg";
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils/tailwindClassConcatenation";
 import HeaderMobileNavigation from "../molecules/HeaderMobileNavigation";
 import { shadcnButtonVariants } from "../shadcn/shadcnButtonStyles";
 ---
