+++ ensawards.org/src/pages/404.astro
@@ -3,7 +3,7 @@ import { Image } from "astro:assets";
 import NotFoundErrorCode from "../assets/404.svg";
 import { shadcnButtonVariants } from "../components/shadcn/shadcnButtonStyles";
 import Layout from "../layouts/Layout.astro";
-import { cn } from "../lib/utils";
+import { cn } from "../utils/tailwindClassConcatenation";
 ---
 
 <Layout isStatic={true}>
