+++ ensawards.org/src/pages/benchmarks/[app].astro
@@ -9,7 +9,8 @@ import SubpageHero from "../../components/molecules/SubpageHero.astro";
 import DataLoadingError from "../../components/organisms/DataLoadingError.astro";
 import { appsData } from "../../data/appData";
 import Layout from "../../layouts/Layout.astro";
-import { BenchmarkResultToBadgeStyles, getAppById, normalizeText } from "../../utils";
+import { getAppById, normalizeText } from "../../utils";
+import {BenchmarkResultToBadgeStyles} from "../../utils/styles";
 
 export function getStaticPaths() {
   return appsData.map((app) => ({ params: { app: normalizeText(app.name) } }));
