+++ ensawards.org/src/utils/index.ts
@@ -24,18 +24,6 @@ export const getSuggestionText = (whatsSuggested: PossibleSuggestions): string =
   }
 };
 
-export const scrollWithOffset = (id: string, offset: number) => {
-  const element = document.getElementById(id);
-  if (!element) return;
-
-  const y = element.getBoundingClientRect().top + window.scrollY - offset;
-
-  window.scrollTo({
-    top: y,
-    behavior: "smooth",
-  });
-};
-
 export const getAppById = (appId: string): AppData | undefined => {
   return appsData.find((app) => app.id === appId);
 };
@@ -56,13 +44,3 @@ export const getBestPracticeByCategoryAndId = (
 
   return category.bestPractices.find((bestPractice) => bestPractice.id === bestPracticeId);
 };
-
-export const getAppSupportColor = (value: number): string =>
-  value > 75 ? "text-emerald-600" : value > 35 ? "text-amber-600" : "text-red-600";
-
-//TODO: distribute these functions more reasonably
-export const BenchmarkResultToBadgeStyles = new Map<BenchmarkResult, string>([
-  [BenchmarkResult.PartialPass, "text-neutral-900 bg-neutral-100"],
-  [BenchmarkResult.Pass, "text-emerald-600 bg-[#0596691A]"],
-  [BenchmarkResult.Fail, "text-red-600 bg-[#DC26261A]"],
-]);
